Fixing rounding with UIColor category

+++ Plaid/UIColor+Utilities.h
@@ -10,7 +10,7 @@
 
 @interface UIColor (Utilities)
 
-+ (UIColor *)colorWithRgbaRed:(int)red green:(int)green blue:(int)blue alpha:(float)alpha;
++ (UIColor *)colorWithRgbaRed:(float)red green:(float)green blue:(float)blue alpha:(float)alpha;
 + (UIColor *)colorWithRgbaRed:(int)red green:(int)green blue:(int)blue;
 
 @end
